+++ app/Http/Controllers/Admin/ManageCollegeInfoController.php
@@ -3,9 +3,9 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
-
 use App\Http\Models\CollegeInfo;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ManageCollegeInfoController extends Controller
 {
@@ -20,6 +20,10 @@ class ManageCollegeInfoController extends Controller
 
     public function collegeInfoUpdate(Request $request,$id)//修改信息
     {
+        //        1.2是学院信息管理权限
+        if(!Auth::user()->can('permission', '1.2'))
+            return response()->view('errors.503');
+
         $collegeInfo=CollegeInfo::find($id);//修改，所以找到对应数据
 
         if ($request->isMethod('post')) {
@@ -49,9 +53,9 @@ class ManageCollegeInfoController extends Controller
             $collegeInfo->college_name = $data['college_name'];
 
             if ($collegeInfo->save() ) {//保持成功与失败
-                return redirect('/admin/manageInfo/college')->with('success', '修改成功!');
+                return redirect('/admin/manageInfo/college')->with('successMsg', '修改成功!');
             } else {
-                return redirect()->back()->with('success', '修改失败!');
+                return redirect()->back()->with('failureMsg', '修改失败!');
             }
         }
         return view('admin.manageInfo.college.update', [//视图
@@ -61,6 +65,10 @@ class ManageCollegeInfoController extends Controller
 
     public function collegeInfoCreate(Request $request)// 新增信息          //    错误信息提示有待于完成
     {
+        //        1.2是学院信息管理权限
+        if(!Auth::user()->can('permission', '1.2'))
+            return response()->view('errors.503');
+
         $collegeInfo=new CollegeInfo();
 
         if ($request->isMethod('post')) {
@@ -86,9 +94,9 @@ class ManageCollegeInfoController extends Controller
             $data = $request->input('CollegeInfo');
 
             if (CollegeInfo::create($data) ) {
-                return redirect('/admin/manageInfo/college')->with('success', '添加成功!');
+                return redirect('/admin/manageInfo/college')->with('successMsg', '添加成功!');
             } else {
-                return redirect()->back();
+                return redirect()->back()->with('failureMsg', '添加失败!');
             }
         }
 
@@ -100,12 +108,16 @@ class ManageCollegeInfoController extends Controller
 
     public function collegeInfoDelete($id)//删除信息
     {
+        //        1.2是学院信息管理权限
+        if(!Auth::user()->can('permission', '1.2'))
+            return response()->view('errors.503');
+
         $collegeInfo=CollegeInfo::find($id);
 
         if($collegeInfo->delete()){
-            return redirect('/admin/manageInfo/college')->with('success', '删除成功!-'.$id);
+            return redirect('/admin/manageInfo/college')->with('successMsg', '删除成功!-'.$id);
         }else {
-            return redirect('/admin/manageInfo/college')->with('error', '删除成功!-'.$id);
+            return redirect('/admin/manageInfo/college')->with('failureMsg', '删除成功!-'.$id);
         }
     }
